+++ MyMusic/MyFan_API/MyFan_API.csproj
@@ -123,7 +123,6 @@
   <ItemGroup>
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Models\JSONResponse.cs" />
-    <Compile Include="Responses\clsFanResult.cs" />
     <Compile Include="Controllers\AlbumsController.cs" />
     <Compile Include="Controllers\BandsController.cs" />
     <Compile Include="Controllers\DiscsController.cs" />
@@ -145,7 +144,7 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="GetForm.cs" />
+    <Compile Include="ControllerCalls\FanControllerCalls.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="packages.config" />
